--- Controls-demo/list_new/RowSeparator/RowSeparatorVisibility/Edges/Edges.wml
@@ -1,7 +1,0 @@
-<div class="controlsDemo__wrapper">
-    <Controls.list:View
-            keyProperty="key"
-            source="{{_viewSource}}"
-            rowSeparatorVisibility="edges"
-            rowSeparatorSize="s"/>
-</div>
