+++ Controls-demo/list_new/RowSeparator/RowSeparatorVisibility/RowSeparatorVisibility.wml
@@ -1,20 +1,12 @@
 <div class="controlsDemo__wrapper">
-    <div class="controlsDemo__cell">
-        <Router.router:Reference attr:class="controls-text-label"
-                                 state="/Controls-demo/app/:app"
-                                 app="{{ 'Controls-demo/list_new/RowSeparator/RowSeparatorVisibility/Edges/Index' }}">
-            <a href="{{content.href}}">Разделители только сверху и снизу</a>
-        </Router.router:Reference>
-        <Controls-demo.list_new.RowSeparator.RowSeparatorVisibility.Edges.Index/>
+    <Controls.list:View
+            keyProperty="key"
+            source="{{_viewSource}}"
+            rowSeparatorVisibility="{{_rowSeparatorVisibility}}"
+            rowSeparatorSize="s"/>
+    <div class="controlsDemo__flexColumn controlsDemo__wrapper__padding-top" style="width: 200px;">
+        <Controls.buttons:Button caption="По умолчанию" on:click="_setRowSeparatorVisibility()"></Controls.buttons:Button>
+        <Controls.buttons:Button caption="Только у элементов" on:click="_setRowSeparatorVisibility('items')"></Controls.buttons:Button>
+        <Controls.buttons:Button caption="Только по краям" on:click="_setRowSeparatorVisibility('edges')"></Controls.buttons:Button>
     </div>
-
-    <div class="controlsDemo__cell">
-        <Router.router:Reference attr:class="controls-text-label"
-                                 state="/Controls-demo/app/:app"
-                                 app="{{ 'Controls-demo/list_new/RowSeparator/RowSeparatorVisibility/Items/Index' }}">
-            <a href="{{content.href}}">Разделители только внктри списка</a>
-        </Router.router:Reference>
-        <Controls-demo.list_new.RowSeparator.RowSeparatorVisibility.Items.Index/>
-    </div>
-
 </div>
